Replace GeoIP MessageFilter with a MessageProcessor
- This makes it possible to change the processor order online.
- Remove the "run before extractors" config, not needed anymore.
- Extract the resolver engine into a separate class.
