+++ src/main/java/org/graylog/plugins/map/MapWidgetModule.java
@@ -16,7 +16,7 @@
  */
 package org.graylog.plugins.map;
 
-import org.graylog.plugins.map.geoip.filter.GeoIpResolverFilter;
+import org.graylog.plugins.map.geoip.processor.GeoIpProcessor;
 import org.graylog.plugins.map.rest.MapDataResource;
 import org.graylog.plugins.map.widget.strategy.MapWidgetStrategy;
 import org.graylog2.plugin.PluginModule;
@@ -24,7 +24,7 @@ import org.graylog2.plugin.PluginModule;
 public class MapWidgetModule extends PluginModule {
     @Override
     protected void configure() {
-        addMessageFilter(GeoIpResolverFilter.class);
+        addMessageProcessor(GeoIpProcessor.class, GeoIpProcessor.Descriptor.class);
         addWidgetStrategy(MapWidgetStrategy.class, MapWidgetStrategy.Factory.class);
         addRestResource(MapDataResource.class);
     }
