+++ src/main/java/org/graylog/plugins/map/config/GeoIpResolverConfig.java
@@ -18,10 +18,12 @@ package org.graylog.plugins.map.config;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.auto.value.AutoValue;
 
 @JsonAutoDetect
+@JsonIgnoreProperties(ignoreUnknown = true)
 @AutoValue
 public abstract class GeoIpResolverConfig {
     public enum DatabaseType {
@@ -37,19 +39,14 @@ public abstract class GeoIpResolverConfig {
     @JsonProperty("db_path")
     public abstract String dbPath();
 
-    @JsonProperty("run_before_extractors")
-    public abstract boolean runBeforeExtractors();
-
     @JsonCreator
     public static GeoIpResolverConfig create(@JsonProperty("enabled") boolean enabled,
                                              @JsonProperty("db_type") DatabaseType dbType,
-                                             @JsonProperty("db_path") String dbPath,
-                                             @JsonProperty("run_before_extractors") boolean runBeforeExtractors) {
+                                             @JsonProperty("db_path") String dbPath) {
         return builder()
                 .enabled(enabled)
                 .dbType(dbType)
                 .dbPath(dbPath)
-                .runBeforeExtractors(runBeforeExtractors)
                 .build();
     }
 
@@ -58,7 +55,6 @@ public abstract class GeoIpResolverConfig {
                .enabled(false)
                .dbType(DatabaseType.MAXMIND_CITY)
                .dbPath("/tmp/GeoLite2-City.mmdb")
-               .runBeforeExtractors(false)
                .build();
     }
 
@@ -73,7 +69,6 @@ public abstract class GeoIpResolverConfig {
         public abstract Builder enabled(boolean enabled);
         public abstract Builder dbType(DatabaseType dbType);
         public abstract Builder dbPath(String dbPath);
-        public abstract Builder runBeforeExtractors(boolean runBeforeExtractors);
 
         public abstract GeoIpResolverConfig build();
     }
