+++ src/main/java/org/graylog/plugins/map/geoip/processor/GeoIpProcessor.java
@@ -1,0 +1,104 @@
+/**
+ * This file is part of Graylog.
+ *
+ * Graylog is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Graylog is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.graylog.plugins.map.geoip.processor;
+
+import com.codahale.metrics.MetricRegistry;
+import com.google.common.eventbus.EventBus;
+import com.google.common.eventbus.Subscribe;
+import org.graylog.plugins.map.config.GeoIpResolverConfig;
+import org.graylog.plugins.map.geoip.GeoIpResolverEngine;
+import org.graylog2.cluster.ClusterConfigChangedEvent;
+import org.graylog2.events.ClusterEventBus;
+import org.graylog2.plugin.Message;
+import org.graylog2.plugin.Messages;
+import org.graylog2.plugin.cluster.ClusterConfigService;
+import org.graylog2.plugin.messageprocessors.MessageProcessor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.inject.Inject;
+import javax.inject.Named;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
+
+public class GeoIpProcessor implements MessageProcessor {
+    private static final Logger LOG = LoggerFactory.getLogger(GeoIpProcessor.class);
+
+    public static class Descriptor implements MessageProcessor.Descriptor {
+        @Override
+        public String name() {
+            return "GeoIP Resolver";
+        }
+
+        @Override
+        public String className() {
+            return GeoIpProcessor.class.getCanonicalName();
+        }
+    }
+
+    private final ClusterConfigService clusterConfigService;
+    private final ScheduledExecutorService scheduler;
+    private final MetricRegistry metricRegistry;
+
+    private final AtomicReference<GeoIpResolverConfig> config;
+    private final AtomicReference<GeoIpResolverEngine> filterEngine;
+
+    @Inject
+    public GeoIpProcessor(ClusterConfigService clusterConfigService,
+                          @Named("daemonScheduler") ScheduledExecutorService scheduler,
+                          @ClusterEventBus EventBus clusterEventBus,
+                          MetricRegistry metricRegistry) {
+        this.clusterConfigService = clusterConfigService;
+        this.scheduler = scheduler;
+        this.metricRegistry = metricRegistry;
+        final GeoIpResolverConfig config = clusterConfigService.getOrDefault(GeoIpResolverConfig.class,
+                GeoIpResolverConfig.defaultConfig());
+
+        this.config = new AtomicReference<>(config);
+        this.filterEngine = new AtomicReference<>(new GeoIpResolverEngine(config, metricRegistry));
+
+        clusterEventBus.register(this);
+    }
+
+    public Messages process(Messages messages) {
+        for (Message message : messages) {
+            filterEngine.get().filter(message);
+        }
+
+        return messages;
+    }
+
+    @Subscribe
+    @SuppressWarnings("unused")
+    public void updateConfig(ClusterConfigChangedEvent event) {
+        if (!GeoIpResolverConfig.class.getCanonicalName().equals(event.type())) {
+            return;
+        }
+
+        scheduler.schedule((Runnable) this::reload, 0, TimeUnit.SECONDS);
+    }
+
+    private void reload() {
+        final GeoIpResolverConfig newConfig = clusterConfigService.getOrDefault(GeoIpResolverConfig.class,
+                GeoIpResolverConfig.defaultConfig());
+
+        LOG.info("Updating GeoIP resolver engine - {}", newConfig);
+        config.set(newConfig);
+        filterEngine.set(new GeoIpResolverEngine(newConfig, metricRegistry));
+    }
+}
